+++ openfisca_us/parameters/irs/credits/child_tax_credit/child/amount.yaml
@@ -16,4 +16,4 @@ metadata:
     - title: 26 U.S. Code § 24(h)
       href: https://www.law.cornell.edu/uscode/text/26/24#h
     - title: 26 U.S. Code § 24(i)
-      href: https://www.law.cornell.edu/uscode/text/26/24#h
+      href: https://www.law.cornell.edu/uscode/text/26/24#i
