+++ openfisca_us/parameters/irs/credits/child_tax_credit/phaseout/arpa/threshold.yaml
@@ -21,4 +21,4 @@ metadata:
   label: CTC phaseout threshold (ARPA)
   reference:
     - title: U.S.C 26 U.S. Code § 24(b)
-      href: https://www.law.cornell.edu/uscode/text/26/24#i
+      href: https://www.law.cornell.edu/uscode/text/26/24#b
