+++ openfisca_us/parameters/irs/credits/child_tax_credit/refundable/phase_in/threshold.yaml
@@ -1,12 +1,14 @@
 description: Additional Child Tax Credit income threshold
 values:
-  2013-01-01: 3_000.0
-  2018-01-01: 2_500.0
-  2026-01-01: 3_000.0
+  2013-01-01: 3_000
+  2018-01-01: 2_500
+  2026-01-01: 3_000
 metadata:
   unit: currency-USD
   name: ctc_refundable_phase_in_threshold
   label: CTC refundable phase-in threshold
   reference:
-    - title: Tax-Calculator
-      href: https://github.com/PSLmodels/Tax-Calculator/blob/6ab24023ed1b661a6dfdd4ce60301838c49536b1/taxcalc/policy_current_law.json#L14333
+    - title: 26 U.S. Code § 24(d)(1)(B)(i)
+      href: https://www.law.cornell.edu/uscode/text/26/24#d_1_B_i
+    - title: 26 U.S. Code § 24(h)(6)
+      href: https://www.law.cornell.edu/uscode/text/26/24#h_6
