+++ openfisca_us/variables/irs/credits/child_tax_credit/reduction.py
@@ -1,8 +1,6 @@
 from openfisca_us.model_api import *
-from openfisca_us.variables.irs.credits.child_tax_credit.maximum import (
-    ctc_child_individual_maximum,
-    ctc_adult_individual_maximum,
-)
+from openfisca_us.variables.irs.credits.child_tax_credit.maximum.individual.child import ctc_child_individual_maximum
+from openfisca_us.variables.irs.credits.child_tax_credit.maximum.individual.adult import ctc_adult_individual_maximum
 from openfisca_us.parameters import default_parameters
 
 
@@ -26,20 +24,29 @@ class ctc_reduction(Variable):
     # TCJA's phaseout changes are purely parametric so don't require structural reform.
 
     def formula_2021(tax_unit, period, parameters):
+        # The ARPA CTC has two phase-outs: the original, and a new phase-out
+        # applying before and only to the increase in the maximum CTC under ARPA.
+
         income = tax_unit("adjusted_gross_income", period)
         ctc = parameters(period).irs.credits.child_tax_credit
         mars = tax_unit("mars", period)
         income_over_threshold = max_(0, income - ctc.phaseout.threshold[mars])
         reduction = ctc.phaseout.rate * income_over_threshold
         maximum_ctc = tax_unit("ctc_maximum", period)
+
+        # Calculate the original phase-out
         original_phaseout = min_(reduction, maximum_ctc)
 
+        # Calculate the income used to assess the new phase-out
         income_over_arpa_threshold = max_(
             0, income - ctc.phaseout.arpa.threshold[mars]
         )
         arpa_phaseout_max_reduction = (
             ctc.phaseout.arpa.rate * income_over_arpa_threshold
         )
+
+        # Calculate the increase - do this by finding the original CTC if
+        # ARPA had not applied - this year's variables, last year's parameters.
         no_arpa_parameters = default_parameters.clone()
         old_ctc = parameters(period.last_year).irs.credits.child_tax_credit
         no_arpa_ctc = no_arpa_parameters.irs.credits.child_tax_credit
@@ -63,6 +70,7 @@ class ctc_reduction(Variable):
             ctc.phaseout.threshold[mars] - ctc.phaseout.arpa.threshold[mars]
         )
 
+        # Apply the phase-out
         arpa_reduction_max = min_(
             arpa_increase,
             ctc.phaseout.arpa.rate * arpa_phaseout_range,
